Fix the app-css task
According to the instructions, the colors.json file should
exist in the views directory root, but this task is actually
reading from the project root dir because gulp automatically
performs a cwd to the gulpfile.js location.

This fix makes the task behavior consistent with the README,
which is presumably ideal.